requirements.txt
annotated-types==0.7.0
appnope==0.1.4
asttokens==2.4.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
conllu==6.0.0
cymem==2.0.8
date-spacy==0.0.1
dateparser==1.2.0
DateRangeParser==1.3.2
dateutils==0.6.12
debugpy==1.8.1
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl
idna==3.10
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.5
jedi==0.19.1
jellyfish==1.1.0
Jinja2==3.1.4
joblib==1.4.2
# Editable install with no version control (journals==0.1)
-e /Users/ben/Projects/LibraryJournals/Code/library-journals
JPype1==1.5.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
langcodes==3.4.1
langdetect==1.0.9
language_data==1.2.0
lockfile==0.12.2
luigi==3.6.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.10
nest-asyncio==1.6.0
nmslib==2.1.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
pendulum==3.0.0
pexpect==4.9.0
platformdirs==4.2.2
preshed==3.0.9
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.2
pysbd==0.3.4
python-daemon==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
RapidFuzz==3.11.0
recordlinkage==0.16
regex==2024.5.15
requests==2.32.3
rich==13.9.2
scikit-learn==1.5.2
scipy==1.10.1
scispacy==0.5.4
setuptools-scm==8.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sutime==1.0.1
tenacity==8.5.0
textdistance==4.6.3
thefuzz==0.22.1
thinc==8.2.5
threadpoolctl==3.5.0
time-machine==2.16.0
timefhuman==0.0.5
tomli==2.0.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wget==3.2
widgetsnbextension==4.0.13
wrapt==1.16.0
xgboost==2.1.3
